fix: Build errors across c6lcd, nrf52 firmware, and dashboard
- Add missing `volatile` qualifier on `mouseState` in C6 state.cpp to match
  common/state.h extern declaration
- Add explicit uint8_t casts in C6 serial_cmd.cpp BMP header to suppress
  narrowing warnings (320 → low byte)
- Add forward declarations for static helpers `dualKeyboardReport()` and
  `trackBleNotify()` in nrf52/hid.cpp (called before definition)
- Remove non-existent `BLESecurity::setSecureConn()` call in ghost_operator.cpp
  (Adafruit Bluefruit API doesn't expose this; SoftDevice handles LESC automatically)
- Install missing vitest dependency in dashboard (was in devDependencies but
  not in node_modules, breaking vite config import)

+++ src/esp32-c6-lcd-1.47/serial_cmd.cpp
@@ -67,8 +67,8 @@ static void takeScreenshot() {
   bmpHdr[4] = fileSize >> 16; bmpHdr[5] = fileSize >> 24;
   bmpHdr[10] = 54;
   bmpHdr[14] = 40;
-  bmpHdr[18] = w; bmpHdr[19] = w >> 8;
-  bmpHdr[22] = h; bmpHdr[23] = h >> 8;
+  bmpHdr[18] = (uint8_t)w; bmpHdr[19] = (uint8_t)(w >> 8);
+  bmpHdr[22] = (uint8_t)h; bmpHdr[23] = (uint8_t)(h >> 8);
   bmpHdr[26] = 1;
   bmpHdr[28] = 24;
   bmpHdr[34] = pixelDataSize; bmpHdr[35] = pixelDataSize >> 8;
